Safari-611.1.10/LayoutTests/fast/scrolling/arrow-key-scroll-in-rtl-document-expected.txt

PASS: scrollLeft is -120
